renaming app --> cli
