+++ setup.py
@@ -2,9 +2,9 @@ from setuptools import setup, find_packages
 import sys
 import os
 
-##########################################################################################
-# HELPER FUNCTIONS #######################################################################
-##########################################################################################
+################################################################################
+# HELPER FUNCTIONS #############################################################
+################################################################################
 
 def get_lookup():
     '''get version by way of singularity.version, returns a 
@@ -58,9 +58,9 @@ LICENSE = lookup['LICENSE']
 with open('README.md') as filey:
     LONG_DESCRIPTION = filey.read()
 
-##########################################################################################
-# MAIN ###################################################################################
-##########################################################################################
+################################################################################
+# MAIN #########################################################################
+################################################################################
 
 
 INSTALL_REQUIRES = get_requirements(lookup)
@@ -100,6 +100,6 @@ setup(
         'Topic :: Utilities'
     ],
 
-    entry_points = {'console_scripts': [ 'som=som.app:main',
+    entry_points = {'console_scripts': [ 'som=som.cli:main',
                                          'som-validator=som.wordfish.cli:main' ] }
 )
