+++ som/cli/static/css/jscolors.css
