+++ som/cli/static/css/mdn-like.css
