+++ som/cli/static/js/dataset_explorer.js
