+++ som/cli/static/js/editor.js
