+++ som/cli/static/js/parsejavascript.js
