+++ som/cli/static/js/sketch.js
