+++ som/cli/static/js/stringstream.js
