+++ som/cli/static/js/tokenize.js
