+++ som/cli/static/js/tokenizejavascript.js
