+++ som/cli/static/js/util.js
